refactor: FastAPI 초기화, YOLO 비동기 처리, DB 자동생성 등 AI 서비스 구조 개선

+++ ppe/orm/violation.py
@@ -1,16 +1,16 @@
-from sqlalchemy import Column, BigInteger, Integer, Text, Float, DateTime, ForeignKey
-from database import Base
+from sqlalchemy import Column, BigInteger, Integer, Text, Float, DateTime, ForeignKey, Double
+from ppe.database import Base
 import datetime
 
 class Violation(Base):
     __tablename__ = "violation"
 
     id = Column(BigInteger, primary_key=True, autoincrement=True)
-    worker_id = Column(BigInteger, ForeignKey("watch.worker_id", ondelete="CASCADE"), nullable=False)
+    worker_id = Column(Integer, ForeignKey("watch.worker_id", ondelete="CASCADE"), nullable=False)
     # ppe_id = Column(BigInteger, nullable=True)
     image_url = Column(Text, nullable=False)
-    latitude = Column(Float, nullable=False)
-    longitude = Column(Float, nullable=False)
+    latitude = Column(Double, nullable=False)
+    longitude = Column(Double, nullable=False)
     timestamp = Column(DateTime, default=datetime.datetime.utcnow, nullable=False)
     helmet_on_count = Column(Integer, default=0)
     seatbelt_on_count = Column(Integer, default=0)
